Updated Excel Spreadsheet Template
Added space for notes for every reading, cleaned up spacing and data typing. Also time is displayed in 24-hour "hhmm" only instead of "hhmmss" as before ("...ss" not really needed). Can make 12-hour AM/PM if there is a request for it, but I personally prefer 24-hour. Also, even though time is displayed in "hhmm", it must still be typed as "hh:mm". Further, for some reason, dates are best entered as int (day number from 1 Jan 1900, I believe...nope it says "0 Jan 1900", but when I enter 30, it says "1 Jan 1900"; go fig). Changed date format from "mmm d" (as "Jan 1") to "dd mmmm" (as "01 January"). It's a 3-month table. All 3 are copies of each other, but a formula can be entered to first day of second and third month to make them continuous.